requirements.txt
adbutils==2.9.4
Appium-Python-Client==5.2.4
attrs==25.3.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
deprecation==2.1.0
Flask==3.1.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
retry2==0.9.5
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
uiautomator2==3.4.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
